requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0
Pillow==10.0.1
svglib==1.5.1
reportlab==4.0.4

# Optional SVG support alternatives (uncomment if active one fails)
cairosvg==2.7.1
# Wand==0.6.13